requirements.txt
# Binance API
python-binance==1.0.19

# Data Analysis
pandas==2.2.2
numpy==1.26.4
ta-lib==0.4.29
ta==0.11.0

# Visualization
matplotlib==3.8.4
plotly==5.20.0
seaborn==0.13.2

# Environment and Configuration
python-dotenv==1.0.1

# Logging
loguru==0.7.2

# HTTP Requests
requests==2.32.3

# Development and Testing
pytest==8.2.2
black==24.4.2
flake8==7.1.0

# Async support
aiohttp==3.9.5
asyncio-throttle==1.0.2

# Utilities
click==8.1.7
rich==13.7.1